better

+++ qa_tina_tests/USER/PERF/perf_objects.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from datetime import datetime
 from string import ascii_lowercase
@@ -5,13 +6,13 @@ from string import ascii_lowercase
 from qa_test_tools import misc
 from qa_tina_tests.USER.PERF.perf_common import log_error
 from qa_tina_tools.tina import oos
-import gc
 
 
 def func_objects(oscsdk, func, service, logger, size, result):
     connector = getattr(oscsdk, service)
     bucket_name = None
     upload_done = False
+    path_to_file = None
     try:
         tmp = misc.id_generator(prefix="bucket", chars=ascii_lowercase)
         connector.create_bucket(Bucket=tmp)
@@ -31,14 +32,14 @@ def func_objects(oscsdk, func, service, logger, size, result):
             mpu_id = mpu.create()
             connector.list_multipart_uploads(Bucket=bucket_name)
             # upload parts
-            logger.debug("beginning of the multipart_upload")
+            logger.debug("beginning of the multipart_upload"+size)
             start_upload = datetime.now()
             parts = mpu.upload(mpu_id)
             upload_duration = (datetime.now() - start_upload).total_seconds()
             upload_done = True
             result["multipart_upload" + service + size] = upload_duration
             logger.debug("end of the multipart_upload")
-            logger.debug("beginning of the list_parts")
+            logger.debug("beginning of the list_parts"+size)
             start_list_parts = datetime.now()
             connector.list_parts(Bucket=bucket_name, Key='data.txt', UploadId=mpu_id)
             list_parts_duration = (datetime.now() - start_list_parts).total_seconds()
@@ -53,25 +54,24 @@ def func_objects(oscsdk, func, service, logger, size, result):
             mpu.complete(mpu_id, parts)
         elif func == 'put_object':
             data = open(path_to_file, "r")
-            logger.debug("beginning of the put_object")
+            logger.debug("beginning of the put_object"+size)
             start_put_object = datetime.now()
             connector.put_object(Bucket=bucket_name, Key='data.txt', Body=str.encode(data.read()))
             put_object_duration = (datetime.now() - start_put_object).total_seconds()
             upload_done = True
             result["put_object" + service + size] = put_object_duration
             logger.debug("end of the put_object")
-            logger.debug("beginning of the get_object")
+            logger.debug("beginning of the get_object"+size)
             start_get_object = datetime.now()
             connector.get_object(Bucket=bucket_name, Key='data.txt')
             get_object_duration = (datetime.now() - start_get_object).total_seconds()
             result["get_object" + service + size] = get_object_duration
             logger.debug("end of the get_object")
-
     except Exception as error:
         log_error(logger, error, "Unexpected error while executing %s".format("bucket_operations"), result)
     finally:
         errors = []
-        if os.path.isfile(path_to_file):
+        if path_to_file and os.path.isfile(path_to_file):
             try:
                 os.remove(path_to_file)
             except Exception as error:
@@ -89,7 +89,6 @@ def func_objects(oscsdk, func, service, logger, size, result):
 
 
 def perf_objects(oscsdk, logger, queue, args):
-
     result = {'status': 'OK'}
     func_objects(oscsdk=oscsdk, func='multipart_upload', service='oos', logger=logger, size='100m', result=result)
     func_objects(oscsdk=oscsdk, func='multipart_upload', service='oos', logger=logger, size='500m', result=result)
@@ -97,13 +96,8 @@ def perf_objects(oscsdk, logger, queue, args):
     func_objects(oscsdk=oscsdk, func='multipart_upload', service='oos', logger=logger, size='2000m', result=result)
     func_objects(oscsdk=oscsdk, func='multipart_upload', service='oos', logger=logger, size='5g', result=result)
     func_objects(oscsdk=oscsdk, func='multipart_upload', service='oos', logger=logger, size='10g', result=result)
-    logger.debug(gc.get_count())
-    gc.collect()
-    logger.debug(gc.get_count())
     func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='100m', result=result)
     func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='500m', result=result)
     func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='1g', result=result)
-    func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='2000m', result=result)
-    func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='5g', result=result)
-    func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='10g', result=result)
+    func_objects(oscsdk=oscsdk, func='put_object', service='oos', logger=logger, size='2g', result=result)
     queue.put(result.copy())
